backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-unfold==0.69.0
djangorestframework==3.16.1
sqlparse==0.5.3
